feat(app-check-client): initialize App Check on web
