+++ .env.production
@@ -8,3 +8,4 @@ VITE_FIREBASE_APP_ID=YOUR_APP_ID
 VITE_FIREBASE_MEASUREMENT_ID=G-XXXXXXXXXX
 VITE_ENABLE_PUBLIC_MARKETING_PAGE=true
 VITE_DEMO_MODE=false
+VITE_RECAPTCHA_V3_SITE_KEY=RECAPTCHA_SITE_KEY_VALUE
